requirements.txt
# Core dependencies
databricks-sdk>=0.18.0
pyspark>=3.4.0
delta-spark>=2.4.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
pyyaml>=6.0.1
click>=8.1.0
rich>=13.6.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Database and storage
sqlalchemy>=2.0.0
pyarrow>=13.0.0

# Utilities
python-dotenv>=1.0.0
croniter>=1.4.0
networkx>=3.1.0
requests>=2.31.0
urllib3>=2.0.0

# Alerting dependencies
slack-sdk>=3.21.0
sendgrid>=6.10.0
pymsteams>=0.2.2
jinja2>=3.1.0

# Monitoring and logging
structlog>=23.1.0
sentry-sdk>=1.32.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
